Convert `ErrorModal` to TSX
